pyproject.toml
[project]
name = "fastapi-skeleton"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asyncpg==0.30.0",
    "bcrypt==4.2.1",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "click==8.1.8",
    "colorama==0.4.6",
    "coverage==7.8.1",
    "croniter==1.4.1",
    "cryptography==44.0.2",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "fastapi==0.115.8",
    "fastapi-cli==0.0.7",
    "fastapi-mail==1.4.2",
    "fastapi-utilities==0.3.1",
    "fastapi-utils==0.8.0",
    "google-generativeai>=0.8.5",
    "greenlet==3.2.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "minio==7.2.15",
    "openai>=1.109.1",
    "packaging==25.0",
    "pdfminer>=20191125",
    "pdfminer-six==20240706",
    "pgvector>=0.4.1",
    "pillow==11.3.0",
    "pluggy==1.6.0",
    "psycopg2-binary==2.9.10",
    "pycparser==2.22",
    "pycryptodome==3.22.0",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pynacl==1.5.0",
    "pyotp==2.9.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-cov==6.0.0",
    "pytest-mock==3.12.0",
    "python-dateutil==2.9.0.post0",
    "python-docx==1.1.2",
    "python-dotenv==1.1.0",
    "python-jose==3.3.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "qrcode==8.2",
    "redis==5.2.1",
    "requests==2.32.5",
    "rsa==4.9.1",
    "sentence-transformers>=5.1.1",
    "sentry-sdk==2.21.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "starlette==0.45.3",
    "typer==0.15.4",
    "typing-extensions==4.13.2",
    "urllib3==2.5.0",
    "uvicorn==0.34.0",
    "watchfiles==1.0.5",
]
